requirements.txt
python-telegram-bot==20.6
pdfplumber==0.10.0
python-dateutil==2.8.0
sqlite3==2.6.0